+++ apps/livreGUI/MainWindow.cpp
@@ -1,4 +1,4 @@
-/* Copyright (c) 2011-2015, EPFL/Blue Brain Project
+/* Copyright (c) 2011-2016, EPFL/Blue Brain Project
  *                          Jafet Villafranca <jafet.villafrancadiaz@epfl.ch>
  *                          Raphael Dumusc <raphael.dumusc@epfl.ch>
  *
@@ -22,8 +22,9 @@
 #include <livreGUI/ui_MainWindow.h>
 
 #include "animationController/AnimationController.h"
+#include "progress/Progress.h"
 #include "renderParametersController/RenderParametersController.h"
-#include "tfEditor/TransferFunctionEditor.h"
+#include "transferFunctionEditor/TransferFunctionEditor.h"
 #include "Controller.h"
 
 #ifdef LIVRE_USE_MONSTEER
@@ -35,20 +36,22 @@ namespace livre
 struct MainWindow::Impl
 {
     Impl( MainWindow* parent, Controller& controller )
-        : _controller( controller )
     {
         _ui.setupUi( parent );
 
         parent->setCentralWidget( new TransferFunctionEditor( controller ));
 
 #ifdef LIVRE_USE_MONSTEER
-        _ui.simulationDockWidget->setWidget( new monsteer::qt::SteeringWidget( ));
+        _ui.simulationDockWidget->setWidget( new monsteer::qt::SteeringWidget );
 #else
         _ui.simulationDockWidget->setHidden( true );
 #endif
 
         _ui.animationDockWidget->setWidget(
-                    new AnimationController( controller ));
+            new AnimationController( controller ));
+
+        _ui.progressDockWidget->setWidget( new Progress( controller ));
+        _ui.progressDockWidget->setHidden( true );
 
         _ui.renderParametersDockWidget->setWidget( new RenderParametersController( controller ));
 
@@ -57,22 +60,18 @@ struct MainWindow::Impl
                                       _ui.renderParametersDockWidget);
     }
 
-    ~Impl()
-    {
-    }
+    ~Impl(){}
 
+private:
     Ui::MainWindow _ui;
-    Controller& _controller;
 };
 
 MainWindow::MainWindow( Controller& controller, QWidget* parent_ )
     : QMainWindow( parent_ )
     , _impl( new Impl( this, controller ))
-{
-}
+{}
 
 MainWindow::~MainWindow()
-{
-}
+{}
 
 }