+++ apps/livreGUI/transferFunctionEditor/TFWidget.cpp
@@ -18,7 +18,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#include <livreGUI/tfEditor/TFWidget.h>
+#include <livreGUI/transferFunctionEditor/TFWidget.h>
 
 #include <QtWidgets>