+++ livre/eq/Channel.cpp
@@ -53,6 +53,10 @@
 #include <livre/core/render/Frustum.h>
 #include <livre/core/render/RenderBrick.h>
 
+#ifdef LIVRE_USE_ZEROEQ
+#  include <zeroeq/publisher.h>
+#endif
+#include <zerobuf/data/Progress.h>
 #include <eq/eq.h>
 #include <eq/gl.h>
 
@@ -68,6 +72,7 @@ public:
           : _channel( channel )
           , _frustum( Matrix4f(), Matrix4f( ))
           , _frameInfo( _frustum, INVALID_FRAME )
+          , _progress( "Loading bricks", 0 )
     {}
 
     void initializeFrame()
@@ -354,6 +359,16 @@ public:
             _channel->drawStatistics();
             drawCacheStatistics();
         }
+
+#ifdef LIVRE_USE_ZEROEQ
+        const size_t all = _frameInfo.allNodes.size();
+        if( all > 0 )
+        {
+            _progress.restart( all );
+            _progress += all - _frameInfo.notAvailableRenderNodes.size();
+            _publisher.publish( _progress );
+        }
+#endif
     }
 
     void drawCacheStatistics()
@@ -588,6 +603,10 @@ public:
     FrameGrabber _frameGrabber;
     FrameInfo _frameInfo;
     std::unique_ptr< RayCastRenderer > _renderer;
+    ::zerobuf::data::Progress _progress;
+#ifdef LIVRE_USE_ZEROEQ
+    zeroeq::Publisher _publisher;
+#endif
 };
 
 Channel::Channel( eq::Window* parent )